SrcPISHING/requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0